Update README.md (#512)
Updated README file. Added a more thorough explanation of the project goal and added a detailed description of how to use the library. I combined the Documentation section with the Usage section and moved the Compiler Guides so that they are next to the Compiler Supports. I also created a longer description about contributions and how to get involved in this project.